refactor(tests): split test_sync.py into a focused sync package (#443)
The monolithic tests/teatree_core/test_sync.py (2325 LOC) is slow to open,
hard to navigate, and discourages adding tests at the right granularity.
Split it verbatim into tests/teatree_core/sync/ by subsystem concern:

- _overlays.py   shared test doubles, fixture MR payloads, mock factories
- conftest.py    the autouse overlay-cache reset that wrapped every test
- test_types.py                value types and pure classify/extract helpers
- test_ticket_extras.py        update_ticket / merge_ticket_extras
- test_review_and_trackers.py  review permalinks, Notion, tracker-404, e2e
- test_terminal_status.py      apply_merged_status / apply_closed_status
- test_followup_core.py        core sync_followup flow
- test_followup_variants.py    assigned issues, work items, merged, labels
- test_github_and_dual.py      _merge_results, dual sync, GitHub backend

Pure test reorganization: zero production change, zero assertion change,
test bodies relocated verbatim with only import paths adjusted. Test
count unchanged (134 before and after, identical class::method set,
134 passed / 0 skipped both before and after).

Relates-to #443
